+++ src/gfx/Item.cpp
@@ -45,7 +45,7 @@ namespace mud
 		for(const ModelItem& item : m_model->m_items)
 		{
 			bgfx::InstanceDataBuffer& buffer = m_instance_buffers[item.m_index];
-			uint32_t num = bgfx::getAvailInstanceDataBuffer(m_instances.size(), sizeof(mat4));
+			uint32_t num = bgfx::getAvailInstanceDataBuffer(uint32_t(m_instances.size()), sizeof(mat4));
 			if(num == 0)
 				return;
 			bgfx::allocInstanceDataBuffer(&buffer, num, sizeof(mat4));
