+++ src/gfx/Texture.cpp
@@ -124,7 +124,7 @@ namespace mud
 
 	void load_texture_rgba(Texture& texture, uint16_t width, uint16_t height, array<uint8_t> data)
 	{
-		const bgfx::Memory* memory = bgfx::alloc(sizeof(uint8_t) * data.m_count);
+		const bgfx::Memory* memory = bgfx::alloc(uint32_t(sizeof(uint8_t) * data.m_count));
 		std::copy(data.m_pointer, data.m_pointer + data.m_count, memory->data);
 		texture.m_texture = bgfx::createTexture2D(width, height, false, 1, bgfx::TextureFormat::RGBA8, GFX_TEXTURE_POINT, memory);
 	}
